requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-dotenv
pyjwt
slowapi